+++ api/app/controllers/api/command_rest.py
@@ -7,7 +7,6 @@ from app.libs import utils
 import json, os
 from app.middlewares.auth import login_required
 from app.helpers import cluster_task
-from time import sleep
 
 class SendCommandRest(Resource):
     def get(self):
@@ -231,7 +230,6 @@ class SendCommandRest(Resource):
 
 
         if init_data['action'] == 'cluster-master':
-            sleep(5)
             result = list()
             for i in init_data['data']:
                 tags = i['tags']
@@ -248,7 +246,6 @@ class SendCommandRest(Resource):
                 return response(200, data=result, message="Master Cluster Processing")
 
         if init_data['action'] == 'cluster-slave':
-            sleep(5)
             result = list()
             for i in init_data['data']:
                 tags = i['tags']